cogs/permissions/requirements.txt
# Custom Permissions System - Dependencies

# Core Discord library
discord.py>=2.0.0

# Unicode text normalization (for fancy Discord names)
unidecode>=1.3.0

# Optional: SQLite async support (for database persistence)
aiosqlite>=0.17.0
